+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - web: Modal show list of uploaded/uploading files
 - web: Modal minimezable
 - web: clear processed files
+- web: current file upload progress ux friendly showing if is done, is uploading of there's an error
 
 ## [0.2.0] - 2022-05-11